// management/frontend/src/api/schemas/cloudStoragePublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface CloudStoragePublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 128
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  provider: string;
  /** @maxLength 512 */
  endpoint?: string;
  /** @maxLength 128 */
  bucket?: string;
  /** @maxLength 64 */
  region?: string;
  /** @minimum 0 */
  used_bytes?: number;
  /** @minimum 0 */
  total_bytes?: number;
  /** @maxLength 32 */
  status?: string;
  is_active?: boolean;
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}